requirements.txt
torch 
torchaudio
torchsummary
matplotlib
transformers
transformers==4.32.0 
accelerate 
tiktoken 
einops 
scipy 
transformers_stream_generator==0.0.4 
peft 
deepspeed
autopep8==1.6.0
Flask==2.1.2
Werkzeug==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.2.0
yapf
pyright
opencc-python-reimplemented
line-bot-sdk
flask_cors
cryptography